Refactoring common Surface methods
