+++ src/TC/src/MemoryInterfaces.cpp
@@ -13,7 +13,7 @@
  * limitations under the License.
  */
 
-#include "MemoryInterfaces.hpp"
+#include "Surfaces.hpp"
 #include <algorithm>
 #include <cstring>
 #include <cuda_runtime.h>
@@ -452,567 +452,3 @@ bool Surface::Empty() const {
 }
 
 CUcontext Surface::Context() { return GetSurfacePlane()->Context(); }
-
-static bool ValidatePlanes(SurfacePlane** planes, size_t planes_num,
-                           size_t surface_elem_size,
-                           size_t surface_num_planes) {
-  if (!planes) {
-    return false;
-  }
-
-  if (planes_num != surface_num_planes) {
-    return false;
-  }
-
-  const auto elem_size = planes[0]->ElemSize();
-  if (surface_elem_size != elem_size) {
-    return false;
-  }
-
-  for (auto i = 0U; i < planes_num; i++) {
-    if (!planes[i]) {
-      return false;
-    }
-
-    if (planes[i]->ElemSize() != elem_size) {
-      return false;
-    }
-  }
-
-  return true;
-}
-
-Surface* SurfaceY::Create() { return new SurfaceY; }
-
-SurfaceY::SurfaceY() {
-  m_planes.clear();
-  m_planes.emplace_back();
-}
-
-SurfaceY::SurfaceY(uint32_t width, uint32_t height, CUcontext context) {
-  m_planes.clear();
-  m_planes.emplace_back(width, height, ElemSize(), DataType(), context);
-}
-
-uint32_t SurfaceY::Width(uint32_t plane) const {
-  return m_planes.at(plane).Width();
-}
-
-uint32_t SurfaceY::WidthInBytes(uint32_t plane) const {
-  return m_planes.at(plane).ElemSize();
-}
-
-uint32_t SurfaceY::Height(uint32_t plane) const {
-  return m_planes.at(plane).Height();
-}
-
-uint32_t SurfaceY::Pitch(uint32_t plane) const {
-  return m_planes.at(plane).Pitch();
-}
-
-CUdeviceptr SurfaceY::PlanePtr(uint32_t plane) {
-  return m_planes.at(plane).GpuMem();
-}
-
-SurfacePlane* SurfaceY::GetSurfacePlane(uint32_t plane) {
-  return &m_planes.at(plane);
-}
-
-bool SurfaceY::Update(SurfacePlane& newPlane) {
-  SurfacePlane* planes[] = {&newPlane};
-  return Update(planes, 1U);
-}
-
-bool SurfaceY::Update(SurfacePlane** pPlanes, size_t planesNum) {
-  if (!ValidatePlanes(pPlanes, planesNum, ElemSize(), 1)) {
-    return false;
-  }
-
-  if (!OwnMemory()) {
-    m_planes.at(0U) = *pPlanes[0];
-    return true;
-  }
-
-  return false;
-}
-
-SurfaceNV12::SurfaceNV12() {
-  m_planes.clear();
-  m_planes.emplace_back();
-}
-
-SurfaceNV12::SurfaceNV12(uint32_t width, uint32_t height, CUcontext context)
-    : SurfaceNV12(width, height, ElemSize(), context) {}
-
-SurfaceNV12::SurfaceNV12(uint32_t width, uint32_t height,
-                         uint32_t hbd_elem_size, CUcontext context) {
-  m_planes.clear();
-  m_planes.emplace_back(width, height * 3 / 2, hbd_elem_size, DataType(),
-                        context);
-}
-
-SurfaceNV12::SurfaceNV12(uint32_t width, uint32_t height, uint32_t pitch,
-                         CUdeviceptr ptr) {
-  auto dlmt_ptr = SurfacePlane::DLPackContext::ToDLPack(
-      width, height * 3 / 2, pitch, ElemSize(), ptr, DataType());
-
-  auto dlmt_smart_ptr =
-      std::shared_ptr<DLManagedTensor>(dlmt_ptr, dlmt_ptr->deleter);
-
-  m_planes.clear();
-  m_planes.emplace_back(*dlmt_smart_ptr.get());
-}
-
-Surface* SurfaceNV12::Create() { return new SurfaceNV12; }
-
-uint32_t SurfaceNV12::Width(uint32_t plane) const {
-  switch (plane) {
-  case 0:
-  case 1:
-    return m_planes.at(0U).Width();
-  default:
-    break;
-  }
-
-  throw invalid_argument("Invalid plane number");
-}
-
-uint32_t SurfaceNV12::WidthInBytes(uint32_t plane) const {
-  return Width(plane) * ElemSize();
-}
-
-uint32_t SurfaceNV12::Height(uint32_t plane) const {
-  switch (plane) {
-  case 0:
-    return m_planes.at(0U).Height() * 2 / 3;
-  case 1:
-    return m_planes.at(0U).Height() / 3;
-  default:
-    break;
-  }
-
-  throw invalid_argument("Invalid plane number");
-}
-
-uint32_t SurfaceNV12::Pitch(uint32_t plane) const {
-  switch (plane) {
-  case 0:
-  case 1:
-    return m_planes.at(0U).Pitch();
-  default:
-    break;
-  }
-
-  throw invalid_argument("Invalid plane number");
-}
-
-CUdeviceptr SurfaceNV12::PlanePtr(uint32_t plane) {
-  if (Empty()) {
-    return 0x0;
-  }
-
-  if (plane < NumPlanes()) {
-    return m_planes.at(0U).GpuMem() + plane * Height() * Pitch();
-  }
-
-  throw invalid_argument("Invalid plane number");
-}
-
-bool SurfaceNV12::Update(SurfacePlane& newPlane) {
-  SurfacePlane* planes[] = {&newPlane};
-  return Update(planes, 1U);
-}
-
-bool SurfaceNV12::Update(SurfacePlane** pPlanes, size_t planesNum) {
-  if (!ValidatePlanes(pPlanes, planesNum, ElemSize(), 1)) {
-    return false;
-  }
-
-  if (!OwnMemory()) {
-    m_planes.at(0U) = *pPlanes[0];
-    return true;
-  }
-
-  return false;
-}
-
-SurfacePlane* SurfaceNV12::GetSurfacePlane(uint32_t plane) {
-  if (plane < NumPlanes()) {
-    return &m_planes.at(0U);
-  }
-
-  return nullptr;
-}
-
-SurfaceP10::SurfaceP10() : SurfaceNV12() {}
-
-SurfaceP10::SurfaceP10(uint32_t width, uint32_t height, CUcontext context)
-    : SurfaceNV12(width, height, ElemSize(), context) {}
-
-Surface* VPF::SurfaceP10::Create() { return new SurfaceP10; }
-
-SurfaceP12::SurfaceP12() : SurfaceNV12() {}
-
-SurfaceP12::SurfaceP12(uint32_t width, uint32_t height, CUcontext context)
-    : SurfaceNV12(width, height, ElemSize(), context) {}
-
-Surface* VPF::SurfaceP12::Create() { return new SurfaceP12; }
-
-SurfaceYUV420::SurfaceYUV420() {
-  m_planes.clear();
-  for (auto i = 0; i < NumPlanes(); i++) {
-    m_planes.emplace_back();
-  }
-}
-
-SurfaceYUV420::SurfaceYUV420(uint32_t width, uint32_t height,
-                             CUcontext context) {
-  m_planes.clear();
-  /* Need to reserve place, otherwise vector may reallocate and SurfacePlane
-   * instances will be copied to new address loosing the memory ownership. Sic!
-   */
-  m_planes.reserve(NumPlanes());
-  // Y plane, full size;
-  m_planes.emplace_back(width, height, ElemSize(), DataType(), context);
-  // U and V planes, decimated size;
-  m_planes.emplace_back(width / 2, height / 2, ElemSize(), DataType(), context);
-  m_planes.emplace_back(width / 2, height / 2, ElemSize(), DataType(), context);
-}
-
-Surface* SurfaceYUV420::Create() { return new SurfaceYUV420; }
-
-uint32_t SurfaceYUV420::Width(uint32_t plane) const {
-  return m_planes.at(plane).Width();
-}
-
-uint32_t SurfaceYUV420::WidthInBytes(uint32_t plane) const {
-  return Width(plane) * ElemSize();
-}
-
-uint32_t SurfaceYUV420::Height(uint32_t plane) const {
-  return m_planes.at(plane).Height();
-}
-
-uint32_t SurfaceYUV420::Pitch(uint32_t plane) const {
-  return m_planes.at(plane).Pitch();
-}
-
-CUdeviceptr SurfaceYUV420::PlanePtr(uint32_t plane) {
-  return m_planes.at(plane).GpuMem();
-}
-
-bool SurfaceYUV420::Update(SurfacePlane& newPlaneY, SurfacePlane& newPlaneU,
-                           SurfacePlane& newPlaneV) {
-  SurfacePlane* planes[] = {&newPlaneY, &newPlaneU, &newPlaneV};
-  return Update(planes, 3);
-}
-
-bool SurfaceYUV420::Update(SurfacePlane** pPlanes, size_t planesNum) {
-  if (OwnMemory() || !ValidatePlanes(pPlanes, planesNum, ElemSize(), 3)) {
-    return false;
-  }
-
-  for (auto i = 0; i < NumPlanes(); i++) {
-    m_planes.at(i) = *pPlanes[i];
-    return true;
-  }
-
-  return false;
-}
-
-SurfacePlane* SurfaceYUV420::GetSurfacePlane(uint32_t plane) {
-  try {
-    return &m_planes.at(plane);
-  } catch (...) {
-    return nullptr;
-  }
-}
-
-SurfaceYUV422::SurfaceYUV422() {
-  m_planes.clear();
-  for (auto i = 0; i < NumPlanes(); i++) {
-    m_planes.emplace_back();
-  }
-}
-
-SurfaceYUV422::SurfaceYUV422(uint32_t width, uint32_t height,
-                             CUcontext context) {
-  m_planes.clear();
-  /* Need to reserve place, otherwise vector may reallocate and SurfacePlane
-   * instances will be copied to new address loosing the memory ownership. Sic!
-   */
-  m_planes.reserve(NumPlanes());
-  // Y plane, full size;
-  m_planes.emplace_back(width, height, ElemSize(), DataType(), context);
-  // U and V planes, decimated size;
-  m_planes.emplace_back(width / 2, height, ElemSize(), DataType(), context);
-  m_planes.emplace_back(width / 2, height, ElemSize(), DataType(), context);
-}
-
-Surface* SurfaceYUV422::Create() { return new SurfaceYUV422; }
-
-uint32_t SurfaceYUV422::Width(uint32_t plane) const {
-  return m_planes.at(plane).Width();
-}
-
-uint32_t SurfaceYUV422::WidthInBytes(uint32_t plane) const {
-  return Width(plane) * ElemSize();
-}
-
-uint32_t SurfaceYUV422::Height(uint32_t plane) const {
-  return m_planes.at(plane).Height();
-}
-
-uint32_t SurfaceYUV422::Pitch(uint32_t plane) const {
-  return m_planes.at(plane).Pitch();
-}
-
-CUdeviceptr SurfaceYUV422::PlanePtr(uint32_t plane) {
-  return m_planes.at(plane).GpuMem();
-}
-
-SurfacePlane* SurfaceYUV422::GetSurfacePlane(uint32_t plane) {
-  try {
-    return &m_planes.at(plane);
-  } catch (...) {
-    return nullptr;
-  }
-}
-
-bool SurfaceYUV422::Update(SurfacePlane& newPlaneY, SurfacePlane& newPlaneU,
-                           SurfacePlane& newPlaneV) {
-  SurfacePlane* planes[] = {&newPlaneY, &newPlaneU, &newPlaneV};
-  return Update(planes, 3);
-}
-
-bool SurfaceYUV422::Update(SurfacePlane** pPlanes, size_t planesNum) {
-  if (OwnMemory() || !ValidatePlanes(pPlanes, planesNum, ElemSize(), 3)) {
-    return false;
-  }
-
-  for (auto i = 0; i < NumPlanes(); i++) {
-    m_planes.at(i) = *pPlanes[i];
-    return true;
-  }
-
-  return false;
-}
-
-SurfaceYUV444::SurfaceYUV444() {
-  m_planes.clear();
-  for (auto i = 0; i < NumPlanes(); i++) {
-    m_planes.emplace_back();
-  }
-}
-
-SurfaceYUV444::SurfaceYUV444(uint32_t width, uint32_t height, CUcontext context)
-    : SurfaceYUV444(width, height, ElemSize(), context) {}
-
-SurfaceYUV444::SurfaceYUV444(uint32_t width, uint32_t height,
-                             uint32_t hbd_elem_size, CUcontext context) {
-  m_planes.clear();
-  /* Need to reserve place, otherwise vector may reallocate and SurfacePlane
-   * instances will be copied to new address loosing the memory ownership. Sic!
-   */
-  m_planes.reserve(NumPlanes());
-  for (auto i = 0; i < NumPlanes(); i++) {
-    m_planes.emplace_back(width, height, hbd_elem_size, DataType(), context);
-  }
-}
-
-Surface* SurfaceYUV444::Create() { return new SurfaceYUV444; }
-
-uint32_t SurfaceYUV444::Width(uint32_t plane) const {
-  return m_planes.at(plane).Width();
-}
-
-uint32_t SurfaceYUV444::WidthInBytes(uint32_t plane) const {
-  return Width(plane) * ElemSize();
-}
-
-uint32_t SurfaceYUV444::Height(uint32_t plane) const {
-  return m_planes.at(plane).Height();
-}
-
-uint32_t SurfaceYUV444::Pitch(uint32_t plane) const {
-  return m_planes.at(plane).Pitch();
-}
-
-CUdeviceptr SurfaceYUV444::PlanePtr(uint32_t plane) {
-  return m_planes.at(plane).GpuMem();
-}
-
-SurfacePlane* SurfaceYUV444::GetSurfacePlane(uint32_t plane) {
-  try {
-    return &m_planes.at(plane);
-  } catch (...) {
-    return nullptr;
-  }
-}
-
-bool SurfaceYUV444::Update(SurfacePlane& newPlaneY, SurfacePlane& newPlaneU,
-                           SurfacePlane& newPlaneV) {
-  SurfacePlane* planes[] = {&newPlaneY, &newPlaneU, &newPlaneV};
-  return Update(planes, 3);
-}
-
-bool SurfaceYUV444::Update(SurfacePlane** pPlanes, size_t planesNum) {
-  if (OwnMemory() || !ValidatePlanes(pPlanes, planesNum, ElemSize(), 3)) {
-    return false;
-  }
-
-  for (auto i = 0; i < NumPlanes(); i++) {
-    m_planes.at(i) = *pPlanes[i];
-    return true;
-  }
-
-  return false;
-}
-
-SurfaceYUV444_10bit::SurfaceYUV444_10bit() : SurfaceYUV444() {}
-
-SurfaceYUV444_10bit::SurfaceYUV444_10bit(uint32_t width, uint32_t height,
-                                         CUcontext context)
-    : SurfaceYUV444(width, height, ElemSize(), context) {}
-
-Surface* VPF::SurfaceYUV444_10bit::Create() { return new SurfaceYUV444_10bit; }
-
-SurfaceRGB::SurfaceRGB() {
-  m_planes.clear();
-  m_planes.emplace_back();
-}
-
-SurfaceRGB::SurfaceRGB(uint32_t width, uint32_t height, CUcontext context)
-    : SurfaceRGB(width, height, ElemSize(), context) {}
-
-SurfaceRGB::SurfaceRGB(uint32_t width, uint32_t height, uint32_t hbd_elem_size,
-                       CUcontext context) {
-  m_planes.clear();
-  m_planes.emplace_back(width * 3, height, hbd_elem_size, DataType(), context);
-}
-
-Surface* SurfaceRGB::Create() { return new SurfaceRGB; }
-
-uint32_t SurfaceRGB::Width(uint32_t plane) const {
-  return m_planes.at(plane).Width() / 3;
-}
-
-uint32_t SurfaceRGB::WidthInBytes(uint32_t plane) const {
-  return m_planes.at(plane).Width() * ElemSize();
-}
-
-uint32_t SurfaceRGB::Height(uint32_t plane) const {
-  return m_planes.at(plane).Height();
-}
-
-uint32_t SurfaceRGB::Pitch(uint32_t plane) const {
-  return m_planes.at(plane).Pitch();
-}
-
-CUdeviceptr SurfaceRGB::PlanePtr(uint32_t plane) {
-  return m_planes.at(plane).GpuMem();
-}
-
-SurfacePlane* SurfaceRGB::GetSurfacePlane(uint32_t plane) {
-  return &m_planes.at(plane);
-}
-
-bool SurfaceRGB::Update(SurfacePlane& newPlane) {
-  SurfacePlane* planes[] = {&newPlane};
-  return Update(planes, 1U);
-}
-
-bool SurfaceRGB::Update(SurfacePlane** pPlanes, size_t planesNum) {
-  if (!ValidatePlanes(pPlanes, planesNum, ElemSize(), 1)) {
-    return false;
-  }
-
-  if (!OwnMemory()) {
-    m_planes.at(0U) = *pPlanes[0];
-    return true;
-  }
-
-  return false;
-}
-
-SurfaceBGR::SurfaceBGR() : SurfaceRGB() {}
-
-SurfaceBGR::SurfaceBGR(uint32_t width, uint32_t height, CUcontext context)
-    : SurfaceRGB(width, height, context) {}
-
-Surface* VPF::SurfaceBGR::Create() { return new SurfaceBGR; }
-
-SurfaceRGB32F::SurfaceRGB32F() : SurfaceRGB() {}
-
-SurfaceRGB32F::SurfaceRGB32F(uint32_t width, uint32_t height, CUcontext context)
-    : SurfaceRGB(width, height, ElemSize(), context) {}
-
-Surface* VPF::SurfaceRGB32F::Create() { return new SurfaceRGB32F; }
-
-SurfaceRGBPlanar::SurfaceRGBPlanar() {
-  m_planes.clear();
-  m_planes.emplace_back();
-}
-
-SurfaceRGBPlanar::SurfaceRGBPlanar(uint32_t width, uint32_t height,
-                                   CUcontext context)
-    : SurfaceRGBPlanar(width, height * 3, ElemSize(), context) {}
-
-VPF::SurfaceRGBPlanar::SurfaceRGBPlanar(uint32_t width, uint32_t height,
-                                        uint32_t hbd_elem_size,
-                                        CUcontext context) {
-  m_planes.clear();
-  m_planes.emplace_back(width, height * 3, hbd_elem_size, DataType(), context);
-}
-
-Surface* SurfaceRGBPlanar::Create() { return new SurfaceRGBPlanar; }
-
-uint32_t SurfaceRGBPlanar::Width(uint32_t plane) const {
-  return m_planes.at(plane).Width();
-}
-
-uint32_t SurfaceRGBPlanar::WidthInBytes(uint32_t plane) const {
-  return Width(plane) * ElemSize();
-}
-
-uint32_t SurfaceRGBPlanar::Height(uint32_t plane) const {
-  return m_planes.at(plane).Height() / 3;
-}
-
-uint32_t SurfaceRGBPlanar::Pitch(uint32_t plane) const {
-  return m_planes.at(plane).Pitch();
-}
-
-CUdeviceptr SurfaceRGBPlanar::PlanePtr(uint32_t plane) {
-  return m_planes.at(plane).GpuMem();
-}
-
-SurfacePlane* SurfaceRGBPlanar::GetSurfacePlane(uint32_t plane) {
-  return &m_planes.at(plane);
-}
-
-bool SurfaceRGBPlanar::Update(SurfacePlane& newPlane) {
-  SurfacePlane* planes[] = {&newPlane};
-  return Update(planes, 1);
-}
-
-bool SurfaceRGBPlanar::Update(SurfacePlane** pPlanes, size_t planesNum) {
-  if (!ValidatePlanes(pPlanes, planesNum, ElemSize(), 1)) {
-    return false;
-  }
-
-  if (!OwnMemory()) {
-    m_planes.at(0U) = *pPlanes[0];
-    return true;
-  }
-
-  return false;
-}
-
-SurfaceRGB32FPlanar::SurfaceRGB32FPlanar() : SurfaceRGBPlanar() {}
-
-SurfaceRGB32FPlanar::SurfaceRGB32FPlanar(uint32_t width, uint32_t height,
-                                         CUcontext context)
-    : SurfaceRGBPlanar(width, height, ElemSize(), context) {}
-
-Surface* VPF::SurfaceRGB32FPlanar::Create() { return new SurfaceRGB32FPlanar; }
